overhaul Quick Start tutorial
Source files were added to automate the tutorial simulations and the
visualization of results (with ParaView and LaTeX PGFPlots).
This way the runs may also be used as a sort of small regression
test (for instance, to check whether a build of SU2 is able to
work correctly).

Moreover, the structure of the tutorial text was reorganized.